pyproject.toml
[project]
name = "foundry-ai-agent-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "azure-ai-projects==1.0.0b10",
    "azure-identity>=1.23.0",
    "click>=8.2.1",
    "openai>=1.88.0",
    "python-dotenv>=1.1.0",
]
